refactor: BestProductList page 폴더로 구조정립> 불필요 주석 정리, styled-components 부분 하단 배치
