+++ src/App.js
@@ -5,13 +5,13 @@ import Nav from "./Components/MainPage/Nav";
 import Footer from "./Components/MainPage/Footer";
 // route
 import Login from "./Components/Login/Login";
-import BestProductList from "./Components/BestProducts/BestProductList";
 import Cart from "./Components/Basket/Cart";
 import AdminMain from "./Components/Admin/AdminMain/Admin";
 import AddProduct from "./Components/Admin/AddProduct/AddProduct";
 import PaymentDetailGuide from "./Components/Payment/PaymentDetailGuide";
 import ProductList from "./Components/ProductList/ProductList";
 import ProductDetail from "./Components/ProductDetail/ProductDetail";
+import BestProductList from "./Pages/BestProducts/BestProductList";
 
 function App() {
   return (
@@ -22,7 +22,7 @@ function App() {
           <Route path="/" element={<MainPage />} />
           <Route path="/cart" element={<Cart />} />
           <Route path="/login" element={<Login />} />
-          <Route path="/bestproducts" element={<BestProductList />} />
+          <Route path="/bestproducts" element={<BestProductList/>} />
           <Route path="/admin" element={<AdminMain />} />
           <Route path="/admin/addproduct" element={<AddProduct />} />
           <Route path="/cart" element={<Cart />} />
