+++ src/Pages/BestProducts/BestProductList.jsx
@@ -1,9 +1,6 @@
 import React, { useEffect, useState } from 'react';
-import bestProducts_data from '../../Data/BestProducts_data.json'
-import BestProducts from './BestProducts';
+import BestProducts from '../../Components/BestProducts/BestProducts';
 import axios from 'axios';
-
-//'http://mini.jh1105.xyz:5305/api/post/normal/?page=0&size=10'
 
 const BestProductList = () => {
 
@@ -13,7 +10,7 @@ const BestProductList = () => {
         const fetchDataFromSpring = async () => {
             try {
                 const response = await axios.get('http://34.192.1.4:8080/api/post/normal/?page=0&size=10');
-                setBestProducts(response.data.content);   //아 이게 원인임.
+                setBestProducts(response.data.content);   
             } catch (error) {
                 console.error('Error fetching data from Spring:', error);
             }
@@ -22,10 +19,8 @@ const BestProductList = () => {
         fetchDataFromSpring();
     }, []);
 
-    //test중
     useEffect(() => {
 
-        console.log('bestProducts 체크 useEffect', bestProducts);
     }, [bestProducts]);
 
     return (
